switched Installation ID from int to long.

+++ Octokit.Reactive/Clients/IObservableAccessTokensClient.cs
@@ -4,6 +4,6 @@ namespace Octokit.Reactive
 {
     public interface IObservableAccessTokensClient
     {
-        IObservable<AccessToken> Create(int installationId);
+        IObservable<AccessToken> Create(long installationId);
     }
 }
